runtime: merge scheduler and scheduler_tpb
