+++ gnuradio-runtime/lib/CMakeLists.txt
@@ -101,7 +101,6 @@ list(APPEND gnuradio_runtime_sources
   realtime.cc
   realtime_impl.cc
   scheduler.cc
-  scheduler_tpb.cc
   sptr_magic.cc
   sync_block.cc
   sync_decimator.cc
